fix: support video playback on native
